requirements.txt
Django==5.1.5
djoser==2.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0